aws-etl/requirements.txt
requests
pandas
boto3
python-dotenv
sqlalchemy
psycopg2-binary